Fixed: modify code style.
