+++ xtest.go
@@ -3,7 +3,6 @@ package main
 import (
 	"fmt"
 	xsfcli "git.iflytek.com/AIaaS/xsf/client"
-	"git.iflytek.com/AIaaS/xsf/utils"
 	"github.com/pterm/pterm"
 	"sync"
 	"time"
@@ -13,31 +12,6 @@ import (
 	"xtest/util"
 	"xtest/var"
 )
-
-func main() {
-	f := _var.NewFlag()
-	f.Parse()
-	// xrpc框架初始化;
-	cli, e := xsfcli.InitClient(_var.CliName, utils.CfgMode(0), utils.WithCfgName(*f.CmdCfg),
-		utils.WithCfgURL(""), utils.WithCfgPrj(""), utils.WithCfgGroup(""),
-		utils.WithCfgService(""), utils.WithCfgVersion(""))
-	if e != nil {
-		fmt.Println("cli xsf init fail with ", e.Error())
-		return
-	}
-
-	// cli配置初始化;
-	conf := _var.NewConf()
-	e = conf.ConfInit(cli.Cfg())
-	if e != nil {
-		fmt.Println("cli conf init fail with ", e.Error())
-		return
-	}
-	//fmt.Printf("%+v\n", conf)
-	x := NewXtest(cli, conf)
-	x.Run()
-	return
-}
 
 type Xtest struct {
 	r   request.Request
@@ -100,7 +74,6 @@ func (x *Xtest) Run() {
 				if x.r.C.LoopCnt.Load() < 0 {
 					break
 				}
-
 
 				switch x.r.C.ReqMode {
 				case 0:
