+++ xtest.toml
@@ -6,8 +6,8 @@ trace-ip = "172.16.51.13"
 service = "AIservice"           # 请求目标服务名称, eg:sms
 svcId = "s12345678"             # 服务id
 timeout = 1000                  # 服务超时时间, 对应服务端waitTime
-multiThr = 5                  # 请求并发数
-loopCnt = 1000                # 请求总次数
+multiThr = 5                    # 请求并发数
+loopCnt = 1000                  # 请求总次数
 sessMode = 0                    # 0: 非会话模式, 1: 常规会话模式 2.文本按行会话模式 3.文件会话模式
 linearMs = 5000                 # 并发压测线性增长区间,单位:ms
 perfOn=true                     # 是否开启性能测试
@@ -15,14 +15,14 @@ perfLevel=0                     # 非会话模式默认0
                                 #会话模式0: 从发第一帧到最后一帧的性能
                                 #会话模式1:首结果(发送第一帧到最后一帧的性能)
                                 #会话模式2:尾结果(发送最后一帧到收到最后一帧的性能)
-service_pid = 227 # AiService 的PID号
-inputCmd = false # jbzhou5 切换为命令行输入，仅在非会话模式生效
-prometheus_switch = true  # jbzhou5 Prometheus开关， 开启后开启双写，同时写入prometheus与本地日志
-prometheus_port = 2117
-plot = true  # jbzhou5 绘制资源图， 默认开启
-plot_file = "./log/line.png"   # jbzhou5 绘制图形保存路径
-file_sorted = 0  # 传入文件是否排序， 0： 随机， 1： 升序， 2： 降序
-file_name_seq = "_" # 传入文件名分割方式 例如传入'_', 则1_2.txt -> 1，2_2.txt -> 2, 为空或者传入非法则不处理
+service_pid = 227               # AiService 的PID号
+inputCmd = false                # jbzhou5 切换为命令行输入，仅在非会话模式生效
+prometheus_switch = true        # jbzhou5 Prometheus开关， 开启后开启双写，同时写入prometheus与本地日志
+prometheus_port = 2117          # jbzhou5 Prometheus指标暴露端口
+plot = true                     # jbzhou5 绘制资源图， 默认开启
+plot_file = "./log/line.png"    # jbzhou5 绘制图形保存路径
+file_sorted = 0                 # 传入文件是否排序， 0： 随机， 1： 升序， 2： 降序
+file_name_seq = "_"             # 传入文件名分割方式 例如传入'_', 则1_2.txt -> 1，2_2.txt -> 2, 为空或者传入非法则不处理
 [header]
 "appid" = "100IME"
 "uid" = "1234567890"
@@ -32,10 +32,10 @@ file_name_seq = "_" # 传入文件名分割方式 例如传入'_', 则1_2.txt ->
 "x" = 1
 
 [data]
-payload = "dataKey2"   # 输入数据流配置段,多个数据流以";"分割， jbzhou5 如果开启了inputCmd， 该值会被清空
+payload = "dataKey2"            # 输入数据流配置段,多个数据流以";"分割， jbzhou5 如果开启了inputCmd， 该值会被清空
 expect = "dataKey3"             # 输出数据流配置段,多个数据流以";"分割
 
-[dataKey1]  # 输入数据流dataKey1描述
+[dataKey1]                      # 输入数据流dataKey1描述
 inputSrc = "path"               # 上行数据流数据源, 配置文件路径(配置为目录则循环读取目录中文件)
 sliceOn = false                 # 切片开关, false:关闭切换, true:开启切片
 sliceSize = 1280                # 上行数据切片大小,用于会话模式: byte
@@ -45,13 +45,13 @@ type = "image"                  # 数据类型，支持"audio","text","image","v
 describe = "k1=v1;k2=v2"        # 数据描述信息,多个描述信息以";"分割
                                 # 图像支持如下属性："encoding", 如"encoding=jpg"
 [dataKey2]
-inputSrc = "./testdata/text2"               # 上行数据流数据源, 配置文件路径(配置为目录则循环读取目录中文件)
+inputSrc = "./testdata/text2"   # 上行数据流数据源, 配置文件路径(配置为目录则循环读取目录中文件)
 sliceOn = false                 # 切片开关, false:关闭切换, true:开启切片
 sliceSize = 1280                # 上行数据切片大小,用于会话模式: byte
 interval = 40                   # 上行数据发包间隔,用于会话模式: ms. 注：动态校准,每个包间隔非严格interval
 name = "input2"                 # 输入数据流key值
-type = "text"                  # 数据类型，支持"audio","text","image","video"
-describe = "encoding=utf8"        # 数据描述信息,多个描述信息以";"分割
+type = "text"                   # 数据类型，支持"audio","text","image","video"
+describe = "encoding=utf8"      # 数据描述信息,多个描述信息以";"分割
                                 # 图像支持如下属性："encoding", 如"encoding=jpg"
 [dataKey3]
 name = "result"                 # 输入数据流key值
@@ -62,7 +62,7 @@ describe = "k1=v1;k2=v2"        # 数据描述信息,多个描述信息以";"分
 
 [downstream]                    # 下行数据流存储输出
 output = 0                      # 输出方式： 0:输出至公共文件outputDst 1:以独立文件形式输出至目录outputDst(文件名:sid+**) 2：输出至终端
-outputDst = "./log/result"          # 响应数据输出目标, 若output=0则配置目录, output=1则配置文件
+outputDst = "./log/result"      # 响应数据输出目标, 若output=0则配置目录, output=1则配置文件
 
 
 [log]
